Phone Controller and Service
Added phone controller,service and searching by different criterias functionalities.

+++ src/main/java/com/cng457/webservice/entity/Phone.java
@@ -15,9 +15,9 @@ public class Phone extends Product {
 
     @Getter
     @Setter
-    private short internalMemory;
+    private int internalMemory;
 
-    public Phone(String brand, String model, String screenSize, Double price, short internalMemory) {
+    public Phone(String brand, String model, String screenSize, Double price, int internalMemory) {
         super();
         setBrand(brand);
         setModel(model);
